+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Releases
 
+- **Added:** Support for compression.zstd (Python 3.14+)
+
 ### v1.3.26 (2026-04-28)
 
 - **Added:** More tests to cover 100% of the codebase.